test(data): share one recording broadcaster across the V4 repository tests
Three test files each carried a private IV4RecordBroadcaster recorder; the
tombstone tests would have been the fourth.
